[Insights]   - Changed CallstackId to 32 bits also for CallstacksProvider API.   - Also fixed static analysis warning (false positive).
#rb Johan.Berg
#preflight 62023f0be85c7a08bbde4499

[CL 18900490 by ionut matasaru in ue5-main branch]

+++ Engine/Source/Developer/TraceServices/Private/Model/CallstacksProvider.h
@@ -20,8 +20,8 @@ public:
 	explicit FCallstacksProvider(IAnalysisSession& Session);
 	virtual ~FCallstacksProvider() {}
 
-	const FCallstack* GetCallstack(uint64 CallstackId) const override;
-	void GetCallstacks(const TArrayView<uint64>& CallstackIds, FCallstack const** OutCallstacks) const override;
+	const FCallstack* GetCallstack(uint32 CallstackId) const override;
+	void GetCallstacks(const TArrayView<uint32>& CallstackIds, FCallstack const** OutCallstacks) const override;
 	void AddCallstack(uint32 CallstackId, const uint64* Frames, uint8 FrameCount);
 
 	// Backward compatibility with legacy memory trace format (5.0-EA).
